CLEANUP: Move bkey/btree constants and macros to bplus_tree
